requirements.txt
absl-py==1.4.0
adabound==0.0.5
appdirs==1.4.4
astor==0.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
autograd==1.6.2
autograd-gamma==0.5.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
blosc2==2.0.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
docker-pycreds==0.4.0
ecos==2.0.12
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.12.2
fonttools==4.40.0
formulaic==0.6.4
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.21.0
google-auth-oauthlib==1.0.0
graphlib-backport==1.0.3
grpcio==1.56.0
h5py==3.7.0
huggingface-hub==0.14.1
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
interface-meta==1.3.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1719845459717/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
joblib==1.3.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257397447/work
kiwisolver==1.4.4
lifelines==0.27.7
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
msgpack==1.0.5
natsort==8.4.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
numexpr==2.8.4
numpy==1.22.4
oauthlib==3.2.2
opencv-python==4.5.1.48
openslide-python==1.2.0
osqp==0.6.3
packaging==23.1
pandas==2.0.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==10.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
protobuf==4.23.4
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyg-lib==0.2.0+pt112cu113
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.0.9
python-dateutil==2.8.2
pytopk==0.1.0
pytz==2023.3
PyYAML==6.0
pyzmq @ file:///croot/pyzmq_1705605076900/work
qdldl==0.1.7.post0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.0.2
scikit-survival==0.17.2
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.40.0
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tables==3.8.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2
threadpoolctl==3.1.0
timm==0.9.2
tokenizers==0.13.3
top-k-models==0.0.2
topk==1.0
torch==1.12.1+cu113
torch-cluster==1.6.1
torch-geometric==2.3.1
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-spline-conv==1.2.2
torchaudio==0.12.1+cu113
torchsummary==1.5.1
torchvision==0.13.1+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.31.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2023.3
urllib3==1.26.16
wandb==0.16.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.0
